2: adding sinh, cosh and a bevy of other functions

+++ include/rpn.h
@@ -21,6 +21,6 @@
 #pragma once
 
 namespace rpn {
-    using machine_t = float;
+    using machine_t = double;
 }
